# web/src/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for the Rack Gateway administration and proxy services.",
        "title": "Rack Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/audit-logs": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Returns paginated audit logs filtered by optional query parameters.",
                "tags": [
                    "Audit"
                ],
                "summary": "List audit logs",
                "parameters": [
                    {
                        "description": "Text search",
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Action type filter",
                        "name": "action_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Resource type filter",
                        "name": "resource_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Status filter",
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Page size",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "ISO8601 start time",
                        "name": "start",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ISO8601 end time",
                        "name": "end",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Relative range (e.g. 24h, 7d, custom)",
                        "name": "range",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by user email",
                        "name": "user",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.AuditLogsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/audit-logs/export": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Streams the filtered audit log dataset as CSV.",
                "tags": [
                    "Audit"
                ],
                "summary": "Export audit logs as CSV",
                "parameters": [
                    {
                        "description": "Text search",
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Action type filter",
                        "name": "action_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Resource type filter",
                        "name": "resource_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Status filter",
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ISO8601 start time",
                        "name": "since",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ISO8601 end time",
                        "name": "until",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "CSRFToken": {
                "description": "HMAC-derived CSRF token tied to the active session.",
                "type": "apiKey",
                "name": "X-CSRF-Token",
                "in": "header"
            },
            "SessionCookie": {
                "description": "HttpOnly session cookie issued after OAuth login.",
                "type": "apiKey",
                "name": "Cookie",
                "in": "header"
            }
        },
        "schemas": {
...
            "db.AuditLog": {
                "type": "object",
                "properties": {
                    "action": {
                        "description": "e.g., \"env.read\", \"user.create\", \"login.oauth_failed\"",
                        "type": "string"
                    },
                    "action_type": {
                        "description": "\"convox\", \"users\", \"auth\"",
                        "type": "string"
                    },
                    "api_token_id": {
                        "type": "integer"
                    },
                    "api_token_name": {
                        "type": "string"
                    },
                    "command": {
                        "type": "string"
                    },
                    "deploy_approval_request_id": {
                        "type": "integer"
                    },
                    "details": {
                        "description": "JSON string",
                        "type": "string"
                    },
                    "event_count": {
                        "type": "integer"
                    },
                    "http_status": {
                        "type": "integer"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "ip_address": {
                        "type": "string"
                    },
                    "rbac_decision": {
                        "description": "\"allow\" or \"deny\"",
                        "type": "string"
                    },
                    "resource": {
                        "type": "string"
                    },
                    "resource_type": {
                        "type": "string"
                    },
                    "response_time_ms": {
                        "type": "integer"
                    },
                    "status": {
                        "description": "\"success\", \"denied\", \"error\", \"blocked\"",
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "user_agent": {
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string"
                    },
                    "user_name": {
                        "type": "string"
                    }
                }
            },
...
            "handlers.AuditLogsResponse": {
                "type": "object",
                "required": [
                    "limit",
                    "logs",
                    "page",
                    "total"
                ],
                "properties": {
                    "limit": {
                        "type": "integer"
                    },
                    "logs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/db.AuditLog"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                }
            },
...
            "handlers.ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}